requirements.txt
requests>=2.25.0
python-dateutil>=2.8.0
pytz>=2021.1